PNDA-3528: Add default queue placement for oozie jobs.
The Oozie jobs don't use the wrapper cli for submitting spark apps but
use the spark client jar directly. For this the queue placement
implemented in the wrapper cli is being circumvented.
The solution here is to define a queue placement in the application.
To assist the developer with this two things have been changed:
1) The default queue assigned to the user will be defined in a property
with key='mapreduce.job.queuename'. This can then be used as value of
the configuration of the 'mapreduce.job.queuename' property.
2) When no config-default.xml is present in the app package, one will be
created for the app with the mapreduce.job.queuename defined. This
configuration is automatically picked up by spark for the queue placement
of the app. This default configuration can still be overruled by
higher priority configurations (ex: defining the
'mapreduce.job.queuename' property in the configuration section of the
workflow definition.